requirements.txt
Flask
Flask-CORS==4.0.0
requests
firebase-admin
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
openai>=1.54.0,<2.0.0
stripe>=6.7.0
PyPDF2==3.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
reportlab
google-search-results
dateparser
aiohttp
asyncpraw
beautifulsoup4
openpyxl>=3.1.0
scikit-learn>=1.3.0
youtube-transcript-api